// src/app/models/student.ts
export interface StudentID{
    sno: number
    name: string;
    course: string;
    
}
export interface StudentList {
    sno: number
    name: string;
    course: string;
    gender: string
    age: number;
    contact: string;
    address: string;
    email: string;
    birthday: string;
    motto: string;
  }